requirements.txt
opencv-python>=4.8
ultralytics>=8.0
deep-sort-realtime>=1.3
pyyaml>=6.0
pytest>=7.0
